Added Maven support, and moved project resources.
This is a large commit that modifies the entire structure of the project. The rational behind making the commit so large was that there is no valid state between "Maven support" and "not Maven support". Support has been added for both a default build with all dependencies included, as well as a release profile build that allows the project to generate a releasable archive containing the program itself and also all the dependencies of the program. This greatly simplifies the build process from the semi-manual process we have used previously.

Part of the Maven support was moving the file structure around to the Maven standardized locations. In addition to moving the files themselves, the loading classes responsible for location of those files have been updated. The only notable change that does not fall exactly under that category was a rename of the "Constants.java" file and class, since the previous name was not really reflecting it's true purpose of housing only resource constants.